[mocks] Converted all the lib tests to use jest mock
