+++ lib/cache.js
@@ -1,25 +1,7 @@
 import { isNull, isArray } from "lodash"
-import { Client } from "memjs"
-import config from "config"
 import { error } from "./loggers"
 
-const { NODE_ENV } = process.env
-const { CACHE_LIFETIME_IN_SECONDS } = config
-
-const createClient = () => {
-  if (NODE_ENV === "test") {
-    const store = {}
-    return {
-      store,
-      get: (key, cb) => cb(null, store[key]),
-      set: (key, data) => (store[key] = data),
-    }
-  }
-  return Client.create(null, {
-    expires: CACHE_LIFETIME_IN_SECONDS,
-  })
-}
-
+import createClient from "lib/loaders/cache/client"
 export const client = createClient()
 
 export default {
