+++ test/lib/cache.js
@@ -1,19 +1,17 @@
-import cache from "lib/cache"
+jest.mock("lib/loaders/cache/client", () => () => ({
+  get: jest.fn(),
+  set: jest.fn(),
+}))
+// import client from "lib/loaders/cache/client"
+
+import cache, { client as internalClient } from "lib/cache"
 
 describe("Cache", () => {
   describe("when connection to Redis fails", () => {
-    beforeAll(() => {
-      cache.__Rewire__("client", {
-        get: (key, cb) => cb(new Error("connect ECONNREFUSED")),
-      })
-    })
-
-    afterAll(() => {
-      cache.__ResetDependency__("client")
-    })
-
     describe("#get", () => {
       it("falls through with a rejection", () => {
+        internalClient.get.mockImplementationOnce((key, cb) => cb(new Error("connect ECONNREFUSED")))
+
         return cache.get("foobar").catch(e => {
           expect(e.message).toEqual("connect ECONNREFUSED")
         })
@@ -22,16 +20,10 @@ describe("Cache", () => {
   })
 
   describe("when successfully connected to the cache", () => {
-    const client = cache.__get__("client")
-
-    afterEach(() => {
-      client.store = {}
-    })
-
     describe("#get", () => {
-      beforeEach(() => cache.set("get_foo", { bar: "baz" }))
-
       it("parses the data and resolves the promise", () => {
+        internalClient.get.mockImplementationOnce((key, cb) => cb(null, JSON.stringify({ bar: "baz" })))
+
         return cache.get("get_foo").then(data => {
           expect(data.bar).toBe("baz")
         })
@@ -39,34 +31,32 @@ describe("Cache", () => {
     })
 
     describe("#set", () => {
+      beforeEach(internalClient.set.mockClear)
+
       describe("with a plain Object", () => {
-        it("sets the cache and includes a timestamp", done => {
+        it("sets the cache and includes a timestamp", () => {
           cache.set("set_foo", { bar: "baz" })
+          expect(internalClient.set).toBeCalledWith("set_foo", expect.stringContaining("cached"), expect.anything())
 
-          client.get("set_foo", (err, data) => {
-            const parsed = JSON.parse(data)
+          const data = internalClient.set.mock.calls[0][1]
+          const parsed = JSON.parse(data)
 
-            expect(parsed.bar).toBe("baz")
-            expect(typeof parsed.cached).toBe("number")
-
-            done()
-          })
+          expect(parsed.bar).toBe("baz")
+          expect(typeof parsed.cached).toBe("number")
         })
       })
 
       describe("with an Array", () => {
-        it("sets the cache and includes a timestamp", done => {
+        it("sets the cache and includes a timestamp", () => {
           cache.set("set_bar", [{ baz: "qux" }])
+          expect(internalClient.set).toBeCalledWith("set_bar", expect.stringContaining("cached"), expect.anything())
 
-          client.get("set_bar", (err, data) => {
-            const parsed = JSON.parse(data)
+          const data = internalClient.set.mock.calls[0][1]
+          const parsed = JSON.parse(data)
 
-            expect(parsed.length).toBe(1)
-            expect(parsed[0].baz).toBe("qux")
-            expect(typeof parsed[0].cached).toBe("number")
-
-            done()
-          })
+          expect(parsed.length).toBe(1)
+          expect(parsed[0].baz).toBe("qux")
+          expect(typeof parsed[0].cached).toBe("number")
         })
       })
     })
